Refactor `patroni_barman_recover` as `patroni_barman` with `recover` sub-command
We currently have a script named `patroni_barman_recover` in Patroni, which is
intended to be used as a custom bootstrap method, or as a custom replica
creation method.

Now there is need of one more Barman related script in Patroni to handle
switching of config models in Barman upon `on_role_change` events.

However, instead of creating another Patroni script, let's say
`patroni_barman_config_switch`, and duplicating a lot of logic in the code,
we decided to refactor the code so:

* Instead of two separate scripts (`patroni_barman_recover` and
  `patroni_barman_config_switch`), we have a single script (`patroni_barman`)
  with 2 sub-commands (`recover` and `config-switch`)

This commit takes care only of the refactoring part, i.e., the `config-switch`
sub-command has not been implemented yet. . This is the overview of changes
that have been performed:

* File `patroni.scripts.barman_recover` has been removed, and its logic
  has been split into a few files:
  * `patroni.scripts.barman.cli`: handles the entrypoint of the new
    `patroni_barman` command, exposing the argument parser and calling
    the appropriate functions depending on the sub-command
  * `patroni.scripts.barman.utils`: implements utilitary enums, functions and
    classes wich can be used by `cli` and by sub-commands implementation:
    * retry mechanism
    * logging set up
    * communication with pg-backup-api
  * `patroni.scripts.barman.recover`: implements the `recover` sub-command only
* File `patroni.tests.test_barman_recover` has been renamed as
  `patroni.tests.test_barman`
* `setup.py` has been changed so it generates a `patroni_barman` application
  instead of `patroni_barman_recover`
* Docs and unit tests were updated accordingly

References: PAT-154.
